refactor(edge/routing): split route_index to decision, host, index, matcher, route , scan, trie, and util modules
