+++ crates/edge/src/routing/mod.rs
@@ -1,0 +1,8 @@
+pub mod decision;
+pub mod host;
+pub mod index;
+pub mod matcher;
+pub mod route;
+pub mod scan;
+pub mod trie;
+pub mod util;
